move Hairdresser to user
