+++ Hairdresser/Program.cs
@@ -40,7 +40,7 @@ builder.Services.AddIdentity<ApplicationUser, IdentityRole>(options =>
                 .AddApiEndpoints();
 
 builder.Services.AddScoped<JWT_Service>();
-builder.Services.AddScoped<IHairdresserService, HairdresserService>();
+builder.Services.AddScoped<IUserService, UserService>();
 builder.Services.AddScoped<IAdminService, AdminService>();
 builder.Services.AddScoped<ITreatmentService, TreatmentService>();
